fix: 배포 에러 해결
- suspense 적용을 위해 로직 분리 및 적용용
